fix(sonar): address minor and major issues

+++ src/main/java/boj/problems/No2261.java
@@ -115,7 +115,7 @@ public class No2261 {
 
     private static long distance(Point a, Point b) {
         long dx = (long) a.x - b.x;
-        long dy = a.y - b.y;
+        long dy = (long) a.y - b.y;
         return dx * dx + dy * dy;
     }
 }
